# docs/openapi.json
{
  "components": {
    "schemas": {
      "AlertState": {
        "enum": [
          "armed",
          "triggered"
        ],
        "title": "AlertState",
        "type": "string"
      },
      "AvailabilityStatus": {
        "enum": [
          "unknown",
          "in_stock",
          "out_of_stock",
          "unavailable"
        ],
        "title": "AvailabilityStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NotificationChannel": {
        "enum": [
          "email",
          "in_app"
        ],
        "title": "NotificationChannel",
        "type": "string"
      },
      "NotificationResponse": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/NotificationChannel"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          }
        },
        "required": [
          "id",
          "channel",
          "status",
          "payload",
          "sent_at",
          "read_at",
          "created_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
      "NotificationStatus": {
        "enum": [
          "pending",
          "sending",
          "sent",
          "failed"
        ],
        "title": "NotificationStatus",
        "type": "string"
      },
      "NotificationUpdate": {
        "properties": {
          "read": {
            "title": "Read",
            "type": "boolean"
          }
        },
        "required": [
          "read"
        ],
        "title": "NotificationUpdate",
        "type": "object"
      },
      "PriceObservationResponse": {
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/AvailabilityStatus"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_price_minor": {
            "title": "Item Price Minor",
            "type": "integer"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "price_minor": {
            "title": "Price Minor",
            "type": "integer"
          },
          "shipping_price_minor": {
            "title": "Shipping Price Minor",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "price_minor",
          "item_price_minor",
          "shipping_price_minor",
          "currency",
          "availability",
          "observed_at"
        ],
        "title": "PriceObservationResponse",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/AvailabilityStatus"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "current_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Minor"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "item_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Price Minor"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "shipping_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Price Minor"
          },
          "store": {
            "$ref": "#/components/schemas/Store"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "store",
          "region",
          "external_id",
          "canonical_url",
          "title",
          "image_url",
          "item_price_minor",
          "shipping_price_minor",
          "current_price_minor",
          "currency",
          "availability",
          "last_checked_at"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
      "Store": {
        "enum": [
          "amazon",
          "ebay"
        ],
        "title": "Store",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchCreate": {
        "properties": {
          "currency": {
            "pattern": "^[A-Za-z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "notify_initial_below_target": {
            "default": true,
            "title": "Notify Initial Below Target",
            "type": "boolean"
          },
          "target_price_minor": {
            "maximum": 9e+18,
            "minimum": 0.0,
            "title": "Target Price Minor",
            "type": "integer"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 8,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "target_price_minor",
          "currency"
        ],
        "title": "WatchCreate",
        "type": "object"
      },
      "WatchResponse": {
        "properties": {
          "alert_state": {
            "$ref": "#/components/schemas/AlertState"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notify_initial_below_target": {
            "title": "Notify Initial Below Target",
            "type": "boolean"
          },
          "product": {
            "$ref": "#/components/schemas/ProductResponse"
          },
          "status": {
            "$ref": "#/components/schemas/WatchStatus"
          },
          "target_price_minor": {
            "title": "Target Price Minor",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_price_minor",
          "currency",
          "status",
          "alert_state",
          "notify_initial_below_target",
          "created_at",
          "updated_at",
          "product"
        ],
        "title": "WatchResponse",
        "type": "object"
      },
      "WatchStatus": {
        "enum": [
          "active",
          "paused"
        ],
        "title": "WatchStatus",
        "type": "string"
      },
      "WatchUpdate": {
        "properties": {
          "notify_initial_below_target": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Initial Below Target"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WatchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_price_minor": {
            "anyOf": [
              {
                "maximum": 9e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price Minor"
          }
        },
        "title": "WatchUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PriceTracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "patch": {
        "operationId": "update_notification_api_v1_notifications__notification_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/watches": {
      "get": {
        "operationId": "list_watches_api_v1_watches_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WatchStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchResponse"
                  },
                  "title": "Response List Watches Api V1 Watches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Watches",
        "tags": [
          "watches"
        ]
      },
      "post": {
        "operationId": "create_watch_api_v1_watches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Watch",
        "tags": [
          "watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}": {
      "delete": {
        "operationId": "delete_watch_api_v1_watches__watch_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Watch",
        "tags": [
          "watches"
        ]
      },
      "get": {
        "operationId": "get_watch_api_v1_watches__watch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Watch",
        "tags": [
          "watches"
        ]
      },
      "patch": {
        "operationId": "update_watch_api_v1_watches__watch_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Watch",
        "tags": [
          "watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/history": {
      "get": {
        "operationId": "watch_history_api_v1_watches__watch_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "30d",
              "enum": [
                "7d",
                "30d",
                "90d",
                "all"
              ],
              "title": "Range",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceObservationResponse"
                  },
                  "title": "Response Watch History Api V1 Watches  Watch Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Watch History",
        "tags": [
          "watches"
        ]
      }
    },
...
  }
}